Move the tests to a separate file
[MAILPOET-3628]
